Update optional fields
In our discussion, status field is supposed to be optional. I have made
the changes in this commit.

In addition, I have also made the comparison on whether an Intern
Application already exists if all of following fields are the same:
* Company Name
* Role
* Cycle

The reason is because, a user may apply for the same company with
different role or at a different cycle.